requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
outcome==1.3.0.post0
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyimgur==0.6.0
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.3
schedule==1.2.2
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
twilio==9.2.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.9.4
